changed overall structure to allow for compatibility with individual so likelihood. Set things up for mflike compatibility only for the moment.
